requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
gspread==5.11.3
oauth2client==4.1.3
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
